+++ lib/component/search_button.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+class SearchButton extends StatelessWidget {
+  const SearchButton({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return ElevatedButton.icon(
+      onPressed: () {},
+      style: ElevatedButton.styleFrom(
+        backgroundColor: Colors.grey[600],
+      ),
+      icon: const Icon(Icons.explore),
+      label: const Text('探索'),
+    );
+  }
+}
